% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiPro.R
\name{actiwatch_ageadjusted}
\alias{actiwatch_ageadjusted}
\title{Convert dataset from ActiWatch to Actipro}
\usage{
actiwatch_ageadjusted(actiwatch_data, age_data_file = NULL)
}
\arguments{
\item{actiwatch_data}{actual actiwatch data using Actiwatch read helpers from ActiPro}

\item{age_data_file}{age data file for varying ages for cutpoints}
}
\value{
A data.table containing the Actipro compatible dataset
}
\description{
Convert dataset from ActiWatch to Actipro
}
