% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiPro.R
\name{actiwatch_mergedcsv_reader}
\alias{actiwatch_mergedcsv_reader}
\title{Convert dataset from ActiWatch XSLX to Actiwatch R}
\usage{
actiwatch_mergedcsv_reader(actiwatch_csv_filepath, drop_suffix = 4,
  id_length = 3)
}
\arguments{
\item{actiwatch_filepath}{filepath of actiwatch data, as a long CSV file}
}
\value{
A data.table containing Actiwatch data, raw
}
\description{
Convert dataset from ActiWatch XSLX to Actiwatch R
}
