% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiPro.R
\name{actiwatch_xlsx_merged_reader}
\alias{actiwatch_xlsx_merged_reader}
\title{Convert dataset from ActiWatch XSLX to Actiwatch R}
\usage{
actiwatch_xlsx_merged_reader(actiwatch_filepath)
}
\arguments{
\item{actiwatch_filepath}{filepath of actiwatch data, as multiple XLSX sheets}
}
\value{
A data.table containing Actiwatch data, raw
}
\description{
Convert dataset from ActiWatch XSLX to Actiwatch R
}
