% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiPro.R
\name{light_features}
\alias{light_features}
\title{Produce novel light activity features}
\usage{
light_features(acc_ageadjusted)
}
\arguments{
\item{acc_ageadjusted}{the file created by ActiPro}
}
\value{
A data.table containing a set of novel statistical parameters based on Keadle
}
\description{
Produce novel light activity features
}
\examples{

}
