% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiPro.R
\name{mvpa_bouts}
\alias{mvpa_bouts}
\title{Generates vector of continuous MVPA bouts}
\usage{
mvpa_bouts(acc_ageadjusted, min_break, act_break, bout_length)
}
\arguments{
\item{acc_ageadjusted}{An ActiPro data.table derived from Actigraph Actilife data}

\item{min_break}{Length of time for a grace period during a bout, probably 0}

\item{act_break}{Grace period bout restriction, upper level}

\item{bout_length}{Minimum bout length, in seconds}
}
\value{
A vector that fits in the \code{acc_ageadjusted} data.table
}
\description{
Generates vector of continuous MVPA bouts
}
\examples{
acc_ageadjusted[, mvpa_bout := function(acc_ageadjusted, 0, 0, 600)

}
