% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiPro.R
\name{mvpa_features}
\alias{mvpa_features}
\title{Produce novel MVPA features}
\usage{
mvpa_features(acc_ageadjusted)
}
\arguments{
\item{acc_ageadjusted}{the file created by ActiPro}
}
\value{
A data.table containing a set of novel statistical parameters based on Keadle
}
\description{
Produce novel MVPA features
}
