% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiPro.R
\name{reach_adapter}
\alias{reach_adapter}
\title{Convert dataset from Reach Lab to Actipro}
\usage{
reach_adapter(rds_filepath)
}
\arguments{
\item{rds_filepath}{file path of rds created by Reach Lab script}
}
\value{
A data.table containing the Actipro compatible dataset
}
\description{
Convert dataset from Reach Lab to Actipro
}
