% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiPro.R
\name{sedentary_feature}
\alias{sedentary_feature}
\alias{sedentary_features}
\title{Produce novel sedentary features}
\usage{
sedentary_features(acc_ageadjusted)
}
\arguments{
\item{acc_ageadjusted}{the file created by ActiPro}
}
\value{
A data.table containing a set of novel statistical parameters based on Keadle
}
\description{
Produce novel sedentary features
}
\examples{

}
