% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiPro.R
\name{acc_ageadjusted}
\alias{acc_ageadjusted}
\title{Produce an activity dataset from raw data}
\usage{
acc_ageadjusted(folder_location, age_data_file, nhanes_nonwear = TRUE,
  id_length = 7, agdFile = FALSE, dataTable = FALSE,
  metaData = TRUE, adult_age_cutpoints = NULL)
}
\arguments{
\item{folder_location}{location of .csv or agd files}

\item{age_data_file}{two column .csv file containing id and age}

\item{nhanes_nonwear}{= TRUE, use NHANES non-wear thresholds}

\item{id_length}{= 7, length of id variable in age_data_file}

\item{agdFile}{= FALSE, agd file, not csv, if TRUE, overrides \code{dataTable} and \code{metaData} settings}

\item{dataTable}{= FALSE, datatable format, post-processed with ActiLife}

\item{metaData}{= TRUE, presence of metadata in headers of csv file}
}
\value{
A data.table containing a variety of activity fields and wear time
}
\description{
Produce an activity dataset from raw data
}
