% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiPro.R
\name{ema_acc_fast}
\alias{ema_acc_fast}
\title{Returns a datatable to merge accelerometer data back into EMA data}
\usage{
ema_acc_fast(ema_file, activity_data, time_stubs = c("15", "30", "60", "120"),
  activity_types = c("VALID", "NONVALID", "MOD", "VIG", "SED", "LIGHT",
  "MVPA", "MVPA_BOUT", "ACTIVITY"))
}
\arguments{
\item{ema_file}{A csv file with three columns, id, fulltime, and index}

\item{activity_data}{An \code{acc_ageadjusted} data.table}

\item{time_stubs}{Time windows to use}

\item{activity_types}{Types of activity from \code{acc_ageadjusted} data.table}
}
\value{
A data.table with windows appended to ema_stubs file
}
\description{
Returns a datatable to merge accelerometer data back into EMA data
}
\examples{
ema_acc_table <- ema_acc[ema_stubs, activity_data]

}
